Externalizing graphing capability
