+++ packages/nexusgraph-graph/package.json
@@ -31,6 +31,8 @@
     "d3-force": "^3.0.0",
     "file-saver": "^2.0.5",
     "lodash": "^4.17.21",
+    "neo4j-devtools-arc": "^0.0.51",
+    "neo4j-driver-core": "5.9.2",
     "re-resizable": "^6.9.9",
     "semantic-ui-react": "^2.1.4"
   },
