--- packages/nexusgraph-graph/src/ArcArrow.test.ts
@@ -1,34 +1,0 @@
-// Copyright 2023 Paion Data. All rights reserved.
-import { ArcArrow } from "./ArcArrow";
-
-let arcArrow: ArcArrow;
-
-beforeEach(() => {
-  arcArrow = new ArcArrow(25, 25, 100, 15, 1, 7, 7, "external");
-});
-
-test("When short Caption Length is 0, the relationship is continuous", () => {
-  expect(arcArrow.outline(0)).toBe(
-    "M 24.018736134675446,6.953439040707552 L 24.277555179777966,5.987513214418485 A 91.67580697480497 " +
-      "91.67580697480497 0 0 0 46.449493255003034,9.098093127373446 L 46.432525782380466,10.0979491694479 A " +
-      "92.67580697480497 92.67580697480497 0 0 1 24.018736134675446,6.953439040707552 Z M " +
-      "46.432525782380466,10.0979491694479 L 46.449493255003034,9.098093127373446 A 91.67580697480497 " +
-      "91.67580697480497 0 0 0 68.71420710757167,6.741599446131563 L 68.03651881255846,3.8191452420759884 L " +
-      "75.64621496620353,5.647402458443353 L 69.6177915009226,10.638205051538995 L 68.9401032059094,7.71575084748342 " +
-      "A 92.67580697480497 92.67580697480497 0 0 1 46.432525782380466,10.0979491694479"
-  );
-});
-
-test("Gets the coordinates of the curve drawn", () => {
-  expect(arcArrow.outline(5)).toBe(
-    "M 24.018736134675446,6.953439040707552 L 24.277555179777966,5.987513214418485 " +
-      "A 91.67580697480497 91.67580697480497 0 0 0 43.964289111225,9.022197931899925 " +
-      "L 43.92021302373678,10.02122610893585 A 92.67580697480497 92.67580697480497 0 0 1 " +
-      "24.018736134675446,6.953439040707552 Z M 48.945995193871425,10.106513208068478 L " +
-      "48.93584157099413,9.106564757425915 A 91.67580697480497 91.67580697480497 0 0 0 " +
-      "68.71420710757167,6.741599446131563 L 68.03651881255846,3.8191452420759884 " +
-      "L 75.64621496620353,5.647402458443353 L 69.6177915009226,10.638205051538995 " +
-      "L 68.9401032059094,7.71575084748342 A 92.67580697480497 92.67580697480497 0 0 1 " +
-      "48.945995193871425,10.106513208068478"
-  );
-});
