--- packages/nexusgraph-graph/src/GraphVisualizer.tsx
@@ -1,126 +1,0 @@
-// Copyright 2023 Paion Data. All rights reserved.
-import { useEffect, useState } from "react";
-
-import { BasicNode, BasicRelationship } from "./basicTypes";
-import { GetNodeNeighboursFn, GraphInteractionCallBack } from "./event-handler";
-import { Graph } from "./Graph";
-import { GraphStats } from "./GraphStats";
-import { GraphStyleModel } from "./GraphStyle";
-import { GraphModel } from "./models/Graph";
-import { RelationshipModel } from "./models/Relationship";
-import { NodesExpandProcessor } from "./processor/NodesExpand";
-import { StyledFullSizeContainer } from "./styles/InspectorContainer.styled";
-import { VizItem } from "./VizItem";
-
-/**
- * Both {@link GraphVisualizerProps.relationships} and {@link GraphVisualizerProps.nodes} are immutable.
- */
-type GraphVisualizerProps = {
-  nodes: readonly BasicNode[];
-  relationships: readonly BasicRelationship[];
-  assignVisElement: (svgElement: any, graphElement: any) => void;
-  onGraphInteraction?: GraphInteractionCallBack;
-};
-
-function computeSelectedItem(nodeLimitHit: boolean, numAllNodes: number, numAllRels: number): VizItem {
-  return nodeLimitHit
-    ? {
-        type: "status-item",
-        item: `Not all return nodes are being displayed due to Initial Node Display setting.
-        Only first ${numAllNodes} nodes are displayed.
-        `,
-      }
-    : {
-        type: "canvas",
-        item: {
-          nodeCount: numAllNodes,
-          relationshipCount: numAllRels,
-        },
-      };
-}
-
-/**
- * Given the provided graph data, {@link GraphVisualizer} turns the data into a graph visualization of the data.
- *
- * {@link GraphVisualizer} is NOT responsible for pre-calculating the graph data; instead it assumes the data has been
- * pre-processed and passed in via {@link GraphVisualizerProps}. {@link GraphVisualizer} guarantees that it does NOT
- * mutate the nodes and relationships passed in it
- *
- * @param props
- * @param styleProps
- * @returns
- */
-export function GraphVisualizer(props: GraphVisualizerProps): JSX.Element {
-  const nodeLimitHit = false;
-  const [selectedItem, setSelectedItem] = useState<VizItem>(
-    computeSelectedItem(nodeLimitHit, props.nodes.length, props.relationships.length)
-  );
-  const [hoveredItem, setHoveredItem] = useState<VizItem>(selectedItem);
-  const [graphStyle, setGraphStyle] = useState<GraphStyleModel>(new GraphStyleModel());
-
-  const [stats, setStats] = useState<GraphStats>({
-    labels: {},
-    relTypes: {},
-  });
-
-  useEffect(() => {
-    setSelectedItem(computeSelectedItem(nodeLimitHit, props.nodes.length, props.relationships.length));
-    setHoveredItem(selectedItem);
-  }, [JSON.stringify(props)]);
-
-  /**
-   * {@link getNodeNeighbours} 在 Neo4J Browser 是借用了 Functioal Programming 的手法，将用户点击“节点展开”的时候被调用，
-   * 调用的大致逻辑是向 Neo4J 数据库发送查询语句，将获取到的数据在这个函数里处理并更新图谱。因为目前我们暂时不支持展开操作，
-   * 但在将来我们需要支持我们自己的展开逻辑（但不是查询 Neo4J，而是我们的后端图谱服务），所以这个方法以一个空方法的方式保留，将来实现 -
-   * https://trello.com/c/59Ypljos
-   *
-   * @param node
-   * @param currentNeighbourIds
-   * @param callback
-   */
-  const getNodeNeighbours: GetNodeNeighboursFn = (node, currentNeighbourIds, callback) => {
-    const nodesExpandProcessor = new NodesExpandProcessor();
-    nodesExpandProcessor.extractionNeighbours(node).then((nodesAndRels) => {
-      callback(nodesAndRels);
-    });
-  };
-
-  return (
-    <StyledFullSizeContainer id="svg-vis">
-      <Graph
-        nodes={props.nodes}
-        relationships={props.relationships}
-        isFullscreen={false}
-        getNodeNeighbours={getNodeNeighbours}
-        onItemMouseOver={(item: VizItem) => setHoveredItem(item)}
-        onItemSelect={(selectedItem: VizItem) => setSelectedItem(selectedItem)}
-        graphStyle={graphStyle}
-        styleVersion={0}
-        onGraphModelChange={(stats) => {
-          setStats(stats);
-        }}
-        // 这个在后续的图谱 PNG 导出需要用到 - https://trello.com/c/GYbX0IEu
-        // 导出逻辑参考 Neo4J Browser CypherFrame.tsx: exportPNG = (): void => { ... }
-        assignVisElement={props.assignVisElement}
-        // 这个在后续的图谱展开需要用到 - 展开出来的新节点不光和被展开的节点有联系，和图上的其它节点也可能存在一种“推算”出来的联系，
-        // 这个方法将执行“推算”逻辑 - https://trello.com/c/KcsSIoK9
-        // Neo4J Browser VisualizationView 的 autoCompleteRelationships 也需要通过某种方式迁移过来，并放在下面的 setGraph
-        autocompleteRelationships={false}
-        getAutoCompleteCallback={(callback: (rels: RelationshipModel[], initialRun: boolean) => void) => {
-          // intentionally left blank
-        }}
-        setGraph={(graph: GraphModel) => {
-          // intentionally left blank
-        }}
-        offset={8}
-        wheelZoomRequiresModKey={true}
-        wheelZoomInfoMessageEnabled={true}
-        disableWheelZoomInfoMessage={() => {
-          // intentionally left blank
-        }}
-        initialZoomToFit={true}
-        onGraphInteraction={props.onGraphInteraction}
-      />
-    </StyledFullSizeContainer>
-  );
-}
