--- packages/nexusgraph-graph/src/LoopArrow.test.ts
@@ -1,30 +1,0 @@
-// Copyright 2023 Paion Data. All rights reserved.
-import { LoopArrow } from "./LoopArrow";
-
-let loopArrow: LoopArrow;
-
-beforeEach(() => {
-  loopArrow = new LoopArrow(10, 40, 30, 4, 10, 10, 6);
-});
-
-test("Gets the coordinates of the curve drawn", () => {
-  expect(loopArrow.outline()).toBe(
-    "M 4.520042103603345 9.141620172685641 L 14.872803907704174 47.77865322424837 " +
-      "A 15.397459621556134 15.397459621556134 0 1 1 -14.872803907704174 47.77865322424837 " +
-      "L -7.108232554628552 18.800878435576323 L -0.34675177060507406 20.61261175129397 " +
-      "L -2.588190451025208 9.659258262890683 L -10.006010033495757 18.024421300268763 " +
-      "L -3.244529249472279 19.836154615986406 L -11.009100602547901 48.81392940465845 " +
-      "A 11.397459621556134 11.397459621556134 0 1 0 11.009100602547901 48.81392940465845 " +
-      "L 0.6563387984470712 10.176896353095724 Z"
-  );
-});
-
-test("Gets the coordinates of the graph covering the curve", () => {
-  expect(loopArrow.overlay(3)).toBe(
-    "M 4.520042103603345 9.141620172685641 L 14.872803907704174 47.77865322424837 " +
-      "A 15.397459621556134 15.397459621556134 0 1 1 -14.872803907704174 47.77865322424837 " +
-      "L -7.108232554628552 18.800878435576323 L -3.244529249472279 19.836154615986406 " +
-      "L -11.009100602547901 48.81392940465845 A 11.397459621556134 11.397459621556134 " +
-      "0 1 0 11.009100602547901 48.81392940465845 L 0.6563387984470712 10.176896353095724 Z"
-  );
-});
