--- packages/nexusgraph-graph/src/event-handler/index.ts
@@ -1,3 +1,0 @@
-// Copyright 2023 Paion Data. All rights reserved.
-export { LINK_ON_CANVAS_CREATE, NODE_ON_CANVAS_CREATE } from "./GraphEventHandlerModel";
-export type { GetNodeNeighboursFn, GraphInteraction, GraphInteractionCallBack } from "./GraphEventHandlerModel";
