--- packages/nexusgraph-graph/src/models/VizItemProperty.ts
@@ -1,6 +1,0 @@
-// Copyright 2023 Paion Data. All rights reserved.
-export interface VizItemProperty {
-  key: string;
-  type: string;
-  value: string;
-}
