--- packages/nexusgraph-graph/src/styles/DefaultPane.styled.ts
@@ -1,80 +1,0 @@
-// Copyright 2023 Paion Data. All rights reserved.
-import styled from "styled-components";
-
-import { StyledLabelChip, StyledRelationshipChip } from "./LabelAndReltypesChip.styled";
-
-export const legendRowHeight = 32;
-
-export const NonClickableLabelChip = styled(StyledLabelChip)`
-  cursor: default;
-`;
-
-export const NonClickableRelTypeChip = styled(StyledRelationshipChip)`
-  cursor: default;
-`;
-
-export const PaneWrapper = styled.div`
-  padding: 0 14px;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  border-style: solid;
-  border-width: 2px;
-  border-color: rgba(130, 221, 248, 0.5);
-  border-radius: 1rem;
-  background-color: transparent;
-`;
-
-export const PaneHeader = styled.div`
-  padding: 5px 0px;
-  font-size: 20px;
-  margin-top: 10px;
-  flex: 0 0 auto;
-  overflow: auto;
-  max-height: 50%;
-  color: #fff;
-  text-shadow: 1px 0.2px 0.2px #fff, 0 0 1em #34c7f4, 0 0 0.2em #34c7f4;
-`;
-
-export const PaneBody = styled.div`
-  height: 100%;
-  overflow: auto;
-  margin: 14px 0;
-  flex: 0 1 auto;
-  display: flex;
-  flex-direction: column;
-  gap: 14px;
-  color: #fff;
-  text-shadow: 1px 0.2px 0.2px #fff, 0 0 1em #34c7f4, 0 0 0.2em #34c7f4;
-`;
-
-export const PaneBodySectionTitle = styled.span`
-  font-weight: 700;
-`;
-
-export const PaneBodySectionSmallText = styled.span`
-  font-size: 0.9rem;
-`;
-export const PaneBodySectionHeaderWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-export const PaneTitle = styled.div`
-  margin-bottom: 10px;
-  display: flex;
-  gap: 5px;
-  align-items: center;
-`;
-
-export const StyledInlineList = styled.ul`
-  list-style: none;
-  word-break: break-word;
-`;
-export const StyledLegendInlineList = styled(StyledInlineList)`
-  padding: 4px 0 0 0;
-  &.contracted {
-    max-height: ${legendRowHeight}px;
-    overflow: hidden;
-  }
-`;
