--- packages/nexusgraph-graph/src/styles/InspectorContainer.styled.ts
@@ -1,65 +1,0 @@
-// Copyright 2023 Paion Data. All rights reserved.
-import styled from "styled-components";
-
-export const panelMinWidth = 200;
-
-export const PaneContainer = styled.div<{
-  paneWidth: number;
-}>`
-  width: ${(props) => props.paneWidth}px;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  background-color: transparent;
-  -webkit-backdrop-filter: blur(8px);
-  backdrop-filter: blur(8px);
-  border-style: solid;
-  border-width: 1px;
-  border-color: rgba(175, 195, 219, 0.5);
-  border-radius: 1rem;
-`;
-
-export const StyledFullSizeContainer = styled.div`
-  position: relative;
-  height: 100%;
-`;
-export const StyledNodeInspectorContainer = styled.div<{
-  paneWidth: number;
-  shouldAnimate: boolean;
-}>`
-  position: absolute;
-  right: 8px;
-  top: 8%;
-  bottom: 8px;
-  width: ${(props) => props.paneWidth}px;
-  ${(props) => props.shouldAnimate && "transition: 0.2s ease-out;"}
-  max-width: 95%;
-  background: ${(props) => props.theme.editorBackground};
-  color: ${(props) => props.theme.primaryText};
-  font-family: ${(props) => props.theme.drawerHeaderFontFamily};
-  box-shadow: ${(props) => props.theme.standardShadow};
-  overflow: hidden;
-`;
-
-export const StyledNodeInspectorTopMenuChevron = styled.button<{
-  expanded: boolean;
-}>`
-  background-color: transparent;
-  border: none;
-  border-radius: 5rem;
-  box-shadow: inset 2px 2px 5px #fff;
-  cursor: pointer;
-  position: absolute;
-  right: 10px;
-  top: 8.9%;
-  width: 32px;
-  height: 32px;
-  padding: 6px;
-  text-align: center;
-  ${(props) =>
-    !props.expanded &&
-    `background: ${props.theme.editorBackground};
-       box-shadow: ${props.theme.standardShadow};
-    `}
-  color: #fff;
-`;
