--- packages/nexusgraph-graph/src/styles/StyledShowMoreButton.styled.ts
@@ -1,9 +1,0 @@
-// Copyright 2023 Paion Data. All rights reserved.
-import styled from "styled-components";
-
-export const StyledShowMoreButton = styled.button`
-  border: none;
-  outline: none;
-  background-color: inherit;
-  color: ${(props) => props.theme.link};
-`;
